added pytest cov and run before
