+++ pytest.ini
@@ -1,0 +1,2 @@
+[pytest]
+python_files=tests/*.py
