+++ test_requirements.txt
@@ -29,6 +29,7 @@ PyInquirer==1.0.3
 PyJWT==1.7.1
 pyparsing==2.4.6
 pytest==5.3.5
+pytest-cov==2.8.1
 PyYAML==5.1.2
 readme-renderer==24.0
 regex==2020.2.20
@@ -38,6 +39,7 @@ smmap==3.0.1
 tabulate==0.8.5
 toml==0.10.0
 tox==3.14.5
+tox-run-before==0.1
 urllib3==1.25.8
 virtualenv==20.0.10
 wcwidth==0.1.8
