+++ tox.ini
@@ -1,7 +1,8 @@
 [tox]
 envlist = py{36, 37}
-;skipsdist = True
 
 [testenv]
-deps = -rrequirements.txt
-commands = pytest
+run_before = pip freeze > test_requirements.txt
+deps = -rtest_requirements.txt
+commands = pytest --cov=src.cobra {posargs}
+skip_install = true
